Implement WeatherService with JooqWeatherRepo, WeatherClient, and Mapper support. Add Weather domain, DTO, exception, and tests. Update pom.xml for forcedTypes.

+++ src/main/java/htwsaar/nordpol/api/dto/WeatherDto.java
@@ -3,9 +3,9 @@ package htwsaar.nordpol.api.dto;
 public record WeatherDto(int session_key,
                          int meeting_key,
                          double air_temperature,
-                         int humidity,
+                         double humidity,
                          int rainfall,
                          double track_temperature,
-                         int wind_direction,
+                         double wind_direction,
                          double wind_speed) {
 }
